Store :max-body in config.edn
Also use "/tmp/" as the default directory to allow for
~$ lein uberjar to run in every environment.

AFAIU the problem is this: wrap-file needs an existing
directory and ~$ lein uberjar will choke at creating the
jar if wrap-file uses a non-existing directory.

Of course, when using ~$ lein run user will want to use
another value in config.edn and when running from a jar
she will probably use environment variables to set this.